feat(httprouterx): refactor helpers
