+++ httprouterx/redir_test.go
@@ -1,4 +1,4 @@
-package x_test
+package httprouterx_test
 
 import (
 	"context"
